# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package db

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Customer struct {
	CustomerUuid           uuid.UUID `json:"customer_uuid"`
	TenantID               string    `json:"tenant_id"`
	CustomerID             string    `json:"customer_id"`
	CustomerName           string    `json:"customer_name"`
	CustomerAddress        string    `json:"customer_address"`
	CustomerCity           string    `json:"customer_city"`
	CustomerState          string    `json:"customer_state"`
	CustomerCountry        string    `json:"customer_country"`
	CustomerTotalValue     int64     `json:"customer_total_value"`
	CustomerStatus         string    `json:"customer_status"`
	CustomerAppType        string    `json:"customer_app_type"`
	CustomerReference      string    `json:"customer_reference"`
	CustomerAppSize        string    `json:"customer_app_size"`
	CustomerPrimaryEmail   string    `json:"customer_primary_email"`
	CustomerPrimaryPhone   string    `json:"customer_primary_phone"`
	CustomerSecondaryEmail string    `json:"customer_secondary_email"`
	CustomerSecondaryPhone string    `json:"customer_secondary_phone"`
}

type CustomerTranscation struct {
	TransactionID    uuid.UUID     `json:"transaction_id"`
	InvoiceID        string        `json:"invoice_id"`
	CustomerUuid     uuid.NullUUID `json:"customer_uuid"`
	TenantID         string        `json:"tenant_id"`
	Amount           sql.NullInt64 `json:"amount"`
	CreatedAt        time.Time     `json:"created_at"`
	UpdatedAt        time.Time     `json:"updated_at"`
	TranscationDebit sql.NullBool  `json:"transcation_debit"`
}

type InternalTranscation struct {
	TransactionID    uuid.UUID     `json:"transaction_id"`
	InvoiceID        string        `json:"invoice_id"`
	UserName         string        `json:"user_name"`
	TenantID         string        `json:"tenant_id"`
	Amount           sql.NullInt64 `json:"amount"`
	CreatedAt        time.Time     `json:"created_at"`
	UpdatedAt        time.Time     `json:"updated_at"`
	TranscationDebit sql.NullBool  `json:"transcation_debit"`
}

type Quotation struct {
	QuotationUuid  uuid.UUID     `json:"quotation_uuid"`
	CustomerUuid   uuid.NullUUID `json:"customer_uuid"`
	QuotationID    string        `json:"quotation_id"`
	TenantID       string        `json:"tenant_id"`
	QuotationValue sql.NullInt64 `json:"quotation_value"`
	QuatationType  string        `json:"quatation_type"`
	CreatedAt      time.Time     `json:"created_at"`
	UpdatedAt      time.Time     `json:"updated_at"`
	Converted      sql.NullBool  `json:"converted"`
}

type ServiceDetail struct {
	CustomerUuid    uuid.NullUUID `json:"customer_uuid"`
	UnderWarranty   sql.NullBool  `json:"under_warranty"`
	NextServiceDate sql.NullTime  `json:"next_service_date"`
	ServiceAmount   sql.NullInt64 `json:"service_amount"`
	TenantID        string        `json:"tenant_id"`
}

type Vendor struct {
	VendorUuid        uuid.UUID `json:"vendor_uuid"`
	VendorName        string    `json:"vendor_name"`
	VendorID          string    `json:"vendor_id"`
	VendorAddress     string    `json:"vendor_address"`
	VendorCity        string    `json:"vendor_city"`
	VendorState       string    `json:"vendor_state"`
	VendorCountry     string    `json:"vendor_country"`
	VendorEmail       string    `json:"vendor_email"`
	VendorPhoneNumber string    `json:"vendor_phone_number"`
	TenantID          string    `json:"tenant_id"`
}

type VendorTranscation struct {
	TransactionID    uuid.UUID     `json:"transaction_id"`
	InvoiceID        string        `json:"invoice_id"`
	VendorUuid       uuid.NullUUID `json:"vendor_uuid"`
	TenantID         string        `json:"tenant_id"`
	Amount           sql.NullInt64 `json:"amount"`
	CreatedAt        time.Time     `json:"created_at"`
	UpdatedAt        time.Time     `json:"updated_at"`
	TransactionType  string        `json:"transaction_type"`
	TranscationDebit sql.NullBool  `json:"transcation_debit"`
}